+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+## 1.0.6
+- Added an example of HTTP endpoint subscription in `topic_example.go`.
+- Added an example of HTTP authorization in `http_authorization.go`.
+- Removed the check for message body size to allow for larger messages.
+
 ## 1.0.5
 - update the minimum Go version declared in go.mod to fix build failures.